Proposed bugfix for heteroskedastic density estimation
Core issue: in some cases, Lrnr_density_semiparametric produces density models that don't have area-under-the-curve = 1.

Diagnosis: The problem is that min_obs_error can get arbitrarily close to 0 (while being positive).  When min_obs_error is used to replace values in the var_preds vector, those are then sqrt()'d and divided by in the `errors / sd_preds` step, leading to the rescaled errors being sent towards +/- Infinity inappropriately.  This causes `approx(density(...))` to yield density curves that don't have area under them approximately equal to 1.

The proposed bugfix is to replace var_preds that are too small with an tolerance based on 10% of the observed standard deviation of the errors from the conditional mean model.